Remove redundant frontend app files and update dependencies
Deleted unused frontend files including HTML, SCSS, and Vite config. Updated dependencies in the `package-lock.json` to reflect the cleanup. This reduces project clutter and ensures a streamlined codebase.

This version is attempting to get rid of boundlers and use pure standard modules when working in development mode with the support of a custom server .